Python: CG trace: Apply better_compare_for_dataclass to all
